store LDO to be sold on the executor's balance
This is due to the limitation of the Aargon's TokenManager which can only assign tokens from its own balance. At the same time we cannot transfer the whole LDO batch to the TokenManager since the latter is a forwarder with anyone owning even 1 LDO wei having the right to execute calls on behalf of TokenManager, so storing any token balance on its address would effectively mean giving that balance away.